@@ -1,17 +1,15 @@
 using System;
 using System.IO;
 using System.Reflection;
-using System.Xml.XPath;
 using Altinn.Platform.Storage.Configuration;
 using Altinn.Platform.Storage.Controllers;
 using Altinn.Platform.Storage.Repository;
-using Halcyon.Web.HAL.Json;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Formatters;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
 
 namespace Altinn.Platform.Storage
 {
@@ -40,6 +38,7 @@ namespace Altinn.Platform.Storage
         /// <param name="services">the service configuration</param>        
         public void ConfigureServices(IServiceCollection services)
         {
+            services.AddControllers().AddNewtonsoftJson();
             services.Configure<AzureCosmosSettings>(Configuration.GetSection("AzureCosmosSettings"));
             services.Configure<AzureStorageConfiguration>(Configuration.GetSection("AzureStorageConfiguration"));
             services.Configure<GeneralSettings>(Configuration.GetSection("BridgeSettings"));
@@ -47,27 +46,21 @@ namespace Altinn.Platform.Storage
             services.AddSingleton<IInstanceRepository, InstanceRepository>();
             services.AddSingleton<IApplicationRepository, ApplicationRepository>();
             services.AddSingleton<IInstanceEventRepository, InstanceEventRepository>();
-
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
-            services.AddMvc().AddControllersAsServices();
-
+            services.AddApplicationInsightsTelemetry();
             services.AddHttpClient<InstancesController>();
 
+            // Need to find out how to add hal with .net core 3.0, is used in IntergrationTest
             // Add HAL support (Halcyon)
-            services.AddMvc().AddMvcOptions(c =>
+            /*services.AddMvc().AddMvcOptions(c =>
             {
-                c.OutputFormatters.RemoveType<JsonOutputFormatter>();
-                c.OutputFormatters.Add(new JsonHalOutputFormatter(new string[] { "application/hal+json" }));
-            });
+                //c.OutputFormatters.RemoveType<NewtonsoftJsonOutputFormatter>();
+                //c.OutputFormatters.Add(new JsonHalOutputFormatter(new string[] { "application/hal+json" }));
+            });*/
 
             // Add Swagger support (Swashbuckle)
             services.AddSwaggerGen(c =>
             {
-                c.SwaggerDoc("v1", new Swashbuckle.AspNetCore.Swagger.Info
-                {
-                    Title = "Altinn Platform Storage",
-                    Version = "v1"
-                });
+                c.SwaggerDoc("v1", new OpenApiInfo { Title = "Altinn Platform Storage", Version = "v1" });
 
                 try
                 {
@@ -75,8 +68,8 @@ namespace Altinn.Platform.Storage
                 }
                 catch
                 {
-                    // Catch swashbuckle exception if it doesn't find the generated XML documentation file
-                }                
+                    // catch swashbuckle exception if it doesn't find the generated xml documentation file
+                }
             });
         }
 
@@ -94,7 +87,7 @@ namespace Altinn.Platform.Storage
         /// </summary>
         /// <param name="app">the application builder</param>
         /// <param name="env">the hosting environment</param>
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -115,7 +108,12 @@ namespace Altinn.Platform.Storage
             });
 
             // app.UseHttpsRedirection();
-            app.UseMvc();
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
