@@ -1,13 +1,13 @@
 using System;
 using System.IO;
 using Microsoft.ApplicationInsights.Extensibility;
-using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Azure.KeyVault;
 using Microsoft.Azure.KeyVault.Models;
 using Microsoft.Azure.Services.AppAuthentication;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Configuration.AzureKeyVault;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Serilog;
 using Serilog.Core;
@@ -30,7 +30,7 @@ namespace Altinn.Platform.Storage
         /// <param name="args">program arguments</param>
         public static void Main(string[] args)
         {
-            CreateWebHostBuilder(args).Build().Run();
+            CreateHostBuilder(args).Build().Run();
         }
 
         /// <summary>
@@ -38,8 +38,12 @@ namespace Altinn.Platform.Storage
         /// </summary>
         /// <param name="args">the arguments</param>
         /// <returns></returns>
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+            .ConfigureWebHostDefaults(webBuilder =>
+            {
+                webBuilder.UseStartup<Startup>();
+            })
             .ConfigureAppConfiguration((hostingContext, config) =>
             {
                 string basePath = Directory.GetParent(Directory.GetCurrentDirectory()).FullName;
@@ -57,15 +61,12 @@ namespace Altinn.Platform.Storage
                                 .CreateLogger();
 
                 logging.AddProvider(new SerilogLoggerProvider(logger));
-            })
-
-            /* // Parameters required for integration with SBL in local development             
-             .UseKestrel()
-             .UseUrls("http://0.0.0.0:5010")
-            .UseContentRoot(Directory.GetCurrentDirectory())
-            .UseIISIntegration() */
-            .UseApplicationInsights()
-            .UseStartup<Startup>();
+            });
+        /* // Parameters required for integration with SBL in local development             
+         .UseKestrel()
+         .UseUrls("http://0.0.0.0:5010")
+        .UseContentRoot(Directory.GetCurrentDirectory())
+        .UseIISIntegration() */
 
         /// <summary>
         /// Load the configuration settings for the program.
